requirement.txt
langchain
langchain_core 
langchain_community 
langgraph 
langchain_groq
python-dotenv
gradio
